requirements.txt
icrawler>=0.6.6
tensorflow>=2.13.0
numpy>=1.24.0
Pillow>=9.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0